Revised title, URL, and abstract.
This appears as having changed everything, which is not quite, but
mostly true. I changed the eol-encoding to Unix and file encoding to
UTF8, which by itself made every line appear to have changed. I did
also revise the abstract in whole, and several fields (though not all)
of the front matter (specifically, title, authors, and URL).

This revision still needs some polishing, in particular adding references.
But it's ready for comments for and review of the draft text itself.